Update README.md with Linus Torvalds' pull request discussion.
This whole repository is really just a means for me to try these new ideas. Once I'm comfortable enough with it I hope I can do the same with my other repositories.

It's not too much to ask to make a good commit message that will make working with the code in the future a lot more easier.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>